requirements.txt
ffmpeg-python>=0.2.0 
pytube>=12.1.0
